refactor-dice-string-parser-create - Moved the Factory out into it's own class.
